+++ examples/App.vue
@@ -1,21 +1,18 @@
 <template>
   <div id="app">
     <input
-      id=""
       v-model="showRow1"
       type="checkbox"
       name="showRow1"
     >
     <label for="showRow1">Show row 1</label>
     <input
-      id=""
-      v-model="showColumn2"
+      v-model="showColumn3"
       type="checkbox"
-      name="showColumn2"
+      name="showColumn3"
     >
-    <label for="showColumn2">Show column 2</label>
+    <label for="showColumn3">Show column 3</label>
     <input
-      id=""
       v-model="toggleSize"
       type="checkbox"
       name="toggle-size"
@@ -31,35 +28,56 @@
       @drag-end="log('drag-end', $event)"
     >
       <SplitGrid
-        v-if="showRow1"
+        :render="showRow1"
         v-bind="splitGridOptions"
+        :strictMode="false"
         class="sb_sub-grid"
         @drag="log('drag', $event)"
         @drag-start="log('drag-start', $event)"
         @drag-end="log('drag-end', $event)"
       >
-        <SplitGridArea>column 1</SplitGridArea>
-        <SplitGridGutter v-if="showColumn2" />
-        <SplitGridArea v-if="showColumn2">
-          column 2
+        <SplitGridArea
+          size-unit="px"
+          :size-value="250"
+        >
+          column 1
         </SplitGridArea>
         <SplitGridGutter />
         <SplitGridArea>
-          column 3
+          column 2
         </SplitGridArea>
+        <SplitGridGutter
+          :show="showColumn3"
+          :transition="transition"
+        />
+        <ReadabilityWrapper :showColumn3="showColumn3"></ReadabilityWrapper>
       </SplitGrid>
-      <SplitGridGutter v-if="showRow1" />
-      <SplitGridArea :size="size">
-        row 3
+      <SplitGridGutter :render="showRow1" />
+      <SplitGridArea
+        :size-unit="size.unit"
+        :size-value="size.value"
+      >
+        row 2
       </SplitGridArea>
     </SplitGrid>
   </div>
 </template>
 <script>
+import ReadabilityWrapper from './components/ReadabilityWrapper';
+
 export default {
   name: 'App',
+  components: {
+    ReadabilityWrapper
+  },
   data() {
     return {
+      showRow1: true,
+      showColumn3: false,
+      toggleSize: false,
+      /**
+       * Vue Split Grid properties
+       */
       animation: {
         duration: 500,
         easing: 'easeInOutQuint'
@@ -68,27 +86,26 @@ export default {
         minSize: 100,
         columnMinSize: 35,
         rowMinSize: 25,
-        columnMinSizes: null,
+        columnMinSizes: { 0: 250 },
         rowMinSizes: null,
         snapOffset: 0,
         columnSnapOffset: 0,
         rowSnapOffset: 0,
         dragInterval: 5,
         columnDragInterval: 5,
-        rowDragInterval: 5,
-        // cursor: null,
-        // columnCursor: null,
-        // rowCursor: null,
-        writeStyle: this.writeStyle
+        rowDragInterval: 5
       },
-      showRow1: true,
-      showColumn2: false,
-      toggleSize: false
+      transition: {
+        name: 'slide-to-left',
+        duration: 200
+      }
     };
   },
   computed: {
     size() {
-      return this.toggleSize ? { value: 500, unit: 'px'} : { value: 50, unit: 'px'};
+      return this.toggleSize
+        ? { value: 500, unit: 'px' }
+        : { value: 50, unit: 'px' };
     }
   },
   methods: {
@@ -107,12 +124,17 @@ body {
 }
 
 #app {
-  font-family: "Avenir", Helvetica, Arial, sans-serif;
+  font-family: 'Avenir', Helvetica, Arial, sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   text-align: center;
   color: #2c3e50;
   height: 100vh;
+  overflow: hidden;
+
+  label {
+    margin-right: 1rem;
+  }
 }
 
 .sb_split-grid {
@@ -120,5 +142,32 @@ body {
 }
 
 .sb_sub-grid {
+}
+
+.fade-enter-active,
+.fade-leave-active {
+  transition: opacity 0.8s;
+}
+.fade-enter,
+.fade-leave-to {
+  opacity: 0;
+}
+.slide-to-right-enter-active,
+.slide-to-right-leave-active {
+  transition: opacity 0.2s, transform 0.2s ease;
+}
+.slide-to-right-enter,
+.slide-to-right-leave-to {
+  opacity: 0;
+  transform: translateX(-20px);
+}
+.slide-to-left-enter-active,
+.slide-to-left-leave-active {
+  transition: opacity 0.2s, transform 0.2s ease;
+}
+.slide-to-left-enter,
+.slide-to-left-leave-to {
+  opacity: 0;
+  transform: translateX(20px);
 }
 </style>