+++ src/mixins/uuid.js
@@ -1,8 +1,5 @@
-let uuid = 0;
-
 export default {
   beforeCreate() {
-    this.uuid = uuid.toString();
-    uuid += 1;
+    this.uuid = String(Math.random()).split('.').join('');
   }
 };